automation/azure/requirements.txt
python-dotenv
azure-identity
azure-core
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network 
azure.mgmt.monitor
azure-eventhub